Create exercise_5_love_calculator.py
This is a mimic of a classic old TRUE LOVE Calculator, where you take the string input of two full names, and count how many letters in both names exist in TRUE, and same for LOVE.  Place the two final numbers side by side, and that's your score.  Output the final score with the True Love results.