Create exercise_1_leap_year_month_count.py
Added to the Leap Year calculator by taking user input of a year and month, and returning the # of days that month has.  Used the leap years stuff because IF that year happens to be a leap year, and the month selected is February, we want to know if it's 28 or 29 days.